refactor(conventions): officially rename error suffix to .errors
This commit renames files not following the `.errors` suffix implicit convention to maintain consistency in naming conventions across the project.
