+++ apps/backend/src/users/authorizations/update-user-role/update-user-role.router.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { createTrpcErrorHandler } from '@tet/backend/utils/trpc/trpc-error-handler';
 import { TrpcService } from '@tet/backend/utils/trpc/trpc.service';
 import z from 'zod';
-import { updateUserRoleErrorConfig } from './update-user-role.error';
+import { updateUserRoleErrorConfig } from './update-user-role.errors';
 import { UpdateUserRoleService } from './update-user-role.service';
 
 const getResultDataOrThrowError = createTrpcErrorHandler(
